+++ firmware/STM32 Drivers/TMC_2590/Drivers/TMC_2590/Src/stm32f1xx_tmc_2590.c
@@ -253,7 +253,7 @@ TMC_2590_StatusTypeDef TMC_2590_MoveSteps(TMC_2590_HandleTypeDef *htmc2590, int1
     return TMC_2590_OK;
 }
 
-TMC_2590_StatusTypeDef TMC_2590_SetTimAutoReload(TMC_2590_HandleTypeDef *htmc2590, uint16_t autoreload)
+TMC_2590_StatusTypeDef TMC_2590_SetTimAutoReload(TMC_2590_HandleTypeDef *htmc2590, uint32_t autoreload)
 {
     __HAL_TIM_SET_AUTORELOAD(htmc2590->Init.STEP_Tim, autoreload);
     return TMC_2590_OK;
